replace dataset classes with AnomalibDataset
